The floats in the compute shader didn't have enough precision so I changed them to doubles

+++ InputHandler.h
@@ -5,9 +5,9 @@
 
 class InputHandler {
 public:
-	float scale;
-	float posx;
-	float posy;
+	double scale;
+	double posx;
+	double posy;
 	bool LMBFirstPress;
 	bool LMBisPressed;
 	double LMBposx;
